RSWW-138 refactor - change location of websocket handlers
