+++ data-generator/src/main/java/cloud/project/datagenerator/hotels/HotelsDataGenerator.java
@@ -6,8 +6,8 @@ import cloud.project.datagenerator.hotels.repositories.HotelRepository;
 import cloud.project.datagenerator.rabbitmq.QueuesConfig;
 import cloud.project.datagenerator.rabbitmq.json.JsonConverter;
 import cloud.project.datagenerator.rabbitmq.requests.RoomUpdateRequest;
-import cloud.project.datagenerator.websockets.DataGeneratorHotelsWebSocketHandler;
-import cloud.project.datagenerator.websockets.HotelUpdate;
+import cloud.project.datagenerator.websockets.hotels.DataGeneratorHotelsWebSocketHandler;
+import cloud.project.datagenerator.websockets.hotels.HotelUpdate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
