+++ data-generator/src/main/java/cloud/project/datagenerator/transports/TransportsDataGenerator.java
@@ -7,8 +7,8 @@ import cloud.project.datagenerator.transports.domain.Transport;
 import cloud.project.datagenerator.transports.domain.TransportCourse;
 import cloud.project.datagenerator.transports.repositories.TransportCourseRepository;
 import cloud.project.datagenerator.transports.repositories.TransportRepository;
-import cloud.project.datagenerator.websockets.DataGeneratorTransportsWebSocketHandler;
-import cloud.project.datagenerator.websockets.TransportUpdate;
+import cloud.project.datagenerator.websockets.transports.DataGeneratorTransportsWebSocketHandler;
+import cloud.project.datagenerator.websockets.transports.TransportUpdate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
